+++ CHANGELOG.md
@@ -11,3 +11,4 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Add sampling handler for sampling records at request level (#3).
 - Add handler to emit JSON logs to GCP Cloud Logging (#6).
+- Add handler to limit records with give rate (#8).